Move serialization classes to Serialization namespace
