+++ modules/platforms/dotnet/Apache.Ignite/Internal/Table/Serialization/TupleSerializerHandler.cs
@@ -15,28 +15,37 @@
  * limitations under the License.
  */
 
-namespace Apache.Ignite.Internal.Table
+namespace Apache.Ignite.Internal.Table.Serialization
 {
-    using System;
     using Buffers;
+    using Ignite.Table;
     using MessagePack;
     using Proto;
 
     /// <summary>
-    /// Object serializer handler.
+    /// Tuple serializer handler.
     /// </summary>
-    /// <typeparam name="T">Object type.</typeparam>
-    internal class ObjectSerializerHandler<T> : IRecordSerializerHandler<T>
-        where T : class
+    internal class TupleSerializerHandler : IRecordSerializerHandler<IIgniteTuple>
     {
-        /// <inheritdoc/>
-        public T Read(ref MessagePackReader reader, Schema schema, bool keyOnly = false)
+        /// <summary>
+        /// Singleton instance.
+        /// </summary>
+        public static readonly TupleSerializerHandler Instance = new();
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="TupleSerializerHandler"/> class.
+        /// </summary>
+        private TupleSerializerHandler()
         {
-            // TODO: Emit code for efficient serialization (IGNITE-16341).
+            // No-op.
+        }
+
+        /// <inheritdoc/>
+        public IIgniteTuple Read(ref MessagePackReader reader, Schema schema, bool keyOnly = false)
+        {
             var columns = schema.Columns;
             var count = keyOnly ? schema.KeyColumnCount : columns.Count;
-            var res = Activator.CreateInstance<T>();
-            var type = typeof(T);
+            var tuple = new IgniteTuple(count);
 
             for (var index = 0; index < count; index++)
             {
@@ -46,45 +55,29 @@ namespace Apache.Ignite.Internal.Table
                 }
 
                 var column = columns[index];
-                var prop = type.GetProperty(column.Name);
-
-                if (prop != null)
-                {
-                    var value = reader.ReadObject(column.Type);
-                    prop.SetValue(value, res);
-                }
-                else
-                {
-                    reader.Skip();
-                }
+                tuple[column.Name] = reader.ReadObject(column.Type);
             }
 
-            return (T)(object)res;
+            return tuple;
         }
 
         /// <inheritdoc/>
-        public T ReadValuePart(PooledBuffer buf, Schema schema, T key)
+        public IIgniteTuple ReadValuePart(PooledBuffer buf, Schema schema, IIgniteTuple key)
         {
-            // TODO: Emit code for efficient serialization (IGNITE-16341).
             // Skip schema version.
             var r = buf.GetReader();
             r.Skip();
 
             var columns = schema.Columns;
-            var res = Activator.CreateInstance<T>();
-            var type = typeof(T);
+            var tuple = new IgniteTuple(columns.Count);
 
             for (var i = 0; i < columns.Count; i++)
             {
                 var column = columns[i];
-                var prop = type.GetProperty(column.Name);
 
                 if (i < schema.KeyColumnCount)
                 {
-                    if (prop != null)
-                    {
-                        prop.SetValue(res, prop.GetValue(key));
-                    }
+                    tuple[column.Name] = key[column.Name];
                 }
                 else
                 {
@@ -93,40 +86,31 @@ namespace Apache.Ignite.Internal.Table
                         continue;
                     }
 
-                    if (prop != null)
-                    {
-                        prop.SetValue(res, r.ReadObject(column.Type));
-                    }
-                    else
-                    {
-                        r.Skip();
-                    }
+                    tuple[column.Name] = r.ReadObject(column.Type);
                 }
             }
 
-            return key;
+            return tuple;
         }
 
         /// <inheritdoc/>
-        public void Write(ref MessagePackWriter writer, Schema schema, T record, bool keyOnly = false)
+        public void Write(ref MessagePackWriter writer, Schema schema, IIgniteTuple record, bool keyOnly = false)
         {
-            // TODO: Emit code for efficient serialization (IGNITE-16341).
             var columns = schema.Columns;
             var count = keyOnly ? schema.KeyColumnCount : columns.Count;
-            var type = record.GetType();
 
             for (var index = 0; index < count; index++)
             {
                 var col = columns[index];
-                var prop = type.GetProperty(col.Name);
+                var colIdx = record.GetOrdinal(col.Name);
 
-                if (prop == null)
+                if (colIdx < 0)
                 {
                     writer.WriteNoValue();
                 }
                 else
                 {
-                    writer.WriteObject(prop.GetValue(record));
+                    writer.WriteObject(record[colIdx]);
                 }
             }
         }
